refactor: rename "user" to "users.types"
