+++ app/users/mongo-user-model.ts
@@ -1,4 +1,4 @@
-import {User, UserStatusOption} from "./user";
+import {User, UserStatusOption} from "./users.types";
 import {Document, Schema} from 'mongoose';
 import database from "../database/mongoose-database";
